Add periodical update and fix post method on city controller.

+++ src/main/java/cz/tul/beran/weather/service/SequenceService.java
@@ -14,7 +14,7 @@ public class SequenceService {
   public Long getNextId() {
     Sequence seq =
         sequenceRepository
-            .findById((long) 1)
+            .findById(1L)
             .orElseGet(
                 () -> {
                   Sequence s = new Sequence();
